+++ job_logger/ui.py
@@ -33,6 +33,8 @@ templates.env.filters["local_display"] = time_utils.format_local_display
 templates.env.filters["local_date"] = time_utils.format_local_date
 templates.env.filters["local_time"] = time_utils.format_local_time
 templates.env.filters["utc_iso"] = time_utils.format_utc_iso
+templates.env.filters["job_date_label"] = time_utils.format_job_date_label
+templates.env.filters["weekday_name"] = time_utils.format_weekday_name
 
 
 @lru_cache(maxsize=1)